Version 0.3
Fixed a bug that caused the calculated future time to be incorrect, off by days.  But, the ccronexpr functions are still not functioning properly; the only commands that works are simple scheduling commands which do not involve turning the wifi both on and off in the same command.  Also, the script does not execute the last scheduled cron command on startup.