Version 0.4
The script works now, for the most part.  The ccronexpr functions still do not work properly using the lua wrapper functions.  So, as a temporary work around, the cron_last_epoch_time_this_line_was_supposedly_executed c program was used in its place and called in a shell command.   And, the script now properly executes the last scheduled cron command on startup.  Unfortunately, all empty lines in the crontab are removed when any modifications are made to the crontab.