Version 0.7
Some bug fixes.  There is no longer a need to configure the script by editing each occurence of the scripts directory after installing, as it now gets the working directory and uses that for the scripts directory in the cron table.  Recurring days of the week now work properly.  The md5 command now includes listing md5 sums of held items.  A couple of new commands have been added, namely 'status wifi' and 'temp', which are self explanatory.